Getting ready for uploading
Disabled ability to click on the company logos since we don’t have
company descriptions for many of the companies.

Also added date to instagram and twitter posts. This is to help
distinguish what is relevant for this years fair. Could also be helped
by adjusting the API query to only get posts from this year.